valve_analytics_webapp/pyproject.toml
[tool.poetry]
name = "valve-analytics-webapp"
version = "0.1.0"
description = ""
authors = ["VladGaSik <vlad_ghaykovich@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-environ = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
